Fix formatting of control limits with user-specfied number of decimals
Previously, trailing zeros were not shown, e.g. with `decimals = 2`
the value 2.5000 would be shown as 2.5 instead of 2.50. This fixes the
problem. Now exactly `decimals` digits are always shown, even if
there are trailing zeros, and even for very large or small numbers
(which would previously be shown using scientific notaiton).